api/requirements-dev.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
Authlib==1.6.9
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cyclonedx-python-lib==11.6.0
defusedxml==0.7.1
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
dparse==0.6.4
drf-spectacular==0.29.0
filelock==3.25.1
flake8==7.3.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.17
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
license-expression==30.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.2
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.1.2
nltk==3.9.3
nodeenv==1.10.0
packageurl-python==0.17.6
packaging==26.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
psycopg==3.3.3
psycopg-binary==3.3.3
py-serializable==2.1.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
pytest-django==4.12.0
python-discovery==1.1.3
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
safety==3.7.0
safety-schemas==0.0.16
setuptools==82.0.1
shellingham==1.5.4
sortedcontainers==2.4.0
sqlparse==0.5.5
tenacity==9.1.4
tomli==2.4.0
tomli_w==1.2.0
tomlkit==0.14.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
virtualenv==21.2.0
wheel==0.46.3
whitenoise==6.12.0
